Fix compile issue

+++ include/chainbase/chainbase.hpp
@@ -477,7 +477,7 @@ namespace chainbase {
             if( _stack.size() != 0 )
                BOOST_THROW_EXCEPTION( std::logic_error("cannot set revision while there is an existing undo stack") );
 
-            if( revision > std::numeric_limits<int64_t>::max() )
+            if( revision > (uint64_t)std::numeric_limits<int64_t>::max() )
                BOOST_THROW_EXCEPTION( std::logic_error("revision to set is too high") );
 
             _revision = static_cast<int64_t>(revision);
